Fixed some compiler warnings and portability issues
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@13148 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ packaging/generic/print-camera-list.c
@@ -1117,8 +1117,8 @@ debug_func (GPLogLevel level, const char *domain, const char *str,
 	struct timeval tv;
 	gettimeofday (&tv,NULL);
 	fprintf (stderr, "%li.%06li %s(%i): %s\n", 
-		 tv.tv_sec - glob_tv_zero.tv_sec, 
-		 (1000000 + tv.tv_usec - glob_tv_zero.tv_usec) % 1000000,
+		 (long) (tv.tv_sec - glob_tv_zero.tv_sec), 
+		 (1000000L + tv.tv_usec - glob_tv_zero.tv_usec) % 1000000L,
 		 domain, level, str);
 }
